fix2

+++ modules/task_1/zaitsev_a_matrix_cannon_block_mult/main.cpp
@@ -4,7 +4,7 @@
 #include <random>
 #include <vector>
 
-std::vector<std::vector<double>> random_matrix(int rows, int columns,
+std::vector<std::vector<double>> random_matrix(size_t rows, size_t columns,
                                                double value) {
   std::vector<std::vector<double>> random(rows, std::vector<double>(columns));
   for (int i = 0; i < rows; i++) {
@@ -17,7 +17,7 @@ std::vector<std::vector<double>> random_matrix(int rows, int columns,
 
 std::vector<std::vector<double>> shift_to_left(
     const std::vector<std::vector<double>>& matrix, int position, int shift,
-    int size) {
+    size_t size) {
   std::vector<double> array(size);
   std::vector<std::vector<double>> shift_to_left = matrix;
   for (int y = 0; y < size; y++) {
@@ -31,7 +31,7 @@ std::vector<std::vector<double>> shift_to_left(
 
 std::vector<std::vector<double>> shift_to_up(
     const std::vector<std::vector<double>>& matrix, int position, int shift,
-    int size) {
+    size_t size) {
   std::vector<double> array(size);
   std::vector<std::vector<double>> shift_to_up = matrix;
   for (int y = 0; y < size; y++) {
@@ -78,16 +78,16 @@ std::vector<std::vector<double>> Cannon_block_matrix_mult(
 
 std::vector<std::vector<double>> matrix_mult(
     const std::vector<std::vector<double>>& first_multiplier,
-    const std::vector<std::vector<double>>& second_multiplier, int size) {
+    const std::vector<std::vector<double>>& second_multiplier, size_t size) {
   std::vector<std::vector<double>> result(size, std::vector<double>(size));
-  for (int i = 0; i < size; i++) {
-    for (int j = 0; j < size; j++) {
+  for (size_t i = 0; i < size; i++) {
+    for (size_t j = 0; j < size; j++) {
       result[i][j] = 0.0;
     }
   }
-  for (int i = 0; i < size; i++) {
-    for (int j = 0; j < size; j++) {
-      for (int k = 0; k < size; k++)
+  for (size_t i = 0; i < size; i++) {
+    for (size_t j = 0; j < size; j++) {
+      for (size_t k = 0; k < size; k++)
         result[i][j] += first_multiplier[i][k] * second_multiplier[k][j];
     }
   }
@@ -101,8 +101,8 @@ TEST(Sequential, Test_1) {
   std::vector<std::vector<double>> C, D;
   D = matrix_mult(A, B, size);
   C = Cannon_block_matrix_mult(A, B, size);
-  for (int i = 0; i < size; i++) {
-    for (int j = 0; j < size; j++) {
+  for (size_t i = 0; i < size; i++) {
+    for (size_t j = 0; j < size; j++) {
       ASSERT_DOUBLE_EQ(C[i][j], D[i][j]);
     }
   }
@@ -115,8 +115,8 @@ TEST(Sequential, Test_2) {
   std::vector<std::vector<double>> C, D;
   D = matrix_mult(A, B, size);
   C = Cannon_block_matrix_mult(A, B, size);
-  for (int i = 0; i < size; i++) {
-    for (int j = 0; j < size; j++) {
+  for (size_t i = 0; i < size; i++) {
+    for (size_t j = 0; j < size; j++) {
       ASSERT_DOUBLE_EQ(C[i][j], D[i][j]);
     }
   }
@@ -129,8 +129,8 @@ TEST(Sequential, Test_3) {
   std::vector<std::vector<double>> C, D;
   D = matrix_mult(A, B, size);
   C = Cannon_block_matrix_mult(A, B, size);
-  for (int i = 0; i < size; i++) {
-    for (int j = 0; j < size; j++) {
+  for (size_t i = 0; i < size; i++) {
+    for (size_t j = 0; j < size; j++) {
       ASSERT_DOUBLE_EQ(C[i][j], D[i][j]);
     }
   }
@@ -143,8 +143,8 @@ TEST(Sequential, Test_4) {
   std::vector<std::vector<double>> C, D;
   D = matrix_mult(A, B, size);
   C = Cannon_block_matrix_mult(A, B, size);
-  for (int i = 0; i < size; i++) {
-    for (int j = 0; j < size; j++) {
+  for (size_t i = 0; i < size; i++) {
+    for (size_t j = 0; j < size; j++) {
       ASSERT_DOUBLE_EQ(C[i][j], D[i][j]);
     }
   }
@@ -157,8 +157,8 @@ TEST(Sequential, Test_5) {
   std::vector<std::vector<double>> C, D;
   D = matrix_mult(A, B, size);
   C = Cannon_block_matrix_mult(A, B, size);
-  for (int i = 0; i < size; i++) {
-    for (int j = 0; j < size; j++) {
+  for (size_t i = 0; i < size; i++) {
+    for (size_t j = 0; j < size; j++) {
       ASSERT_DOUBLE_EQ(C[i][j], D[i][j]);
     }
   }
